Fix TDNF build configuration to use pkgconfig data to find dependencies
With the notable exception of identifying that we're using rpm.org rpm,
all the configuration of how to use dependencies to compile and link
TDNF and its libraries should come from pkgconfig. That way, TDNF
will reliably build across more platforms.

In addition, trivial changes to drop -Werror and specify that -std=gnu99
is required for the build are included so that it will compile properly
on compilers that do not default to newer C standards and not make
assumptions about what is coerced from a warning to an error by default.